Sequencer: Fix rounding error.
[CL 2634370 by Max Chen in Main branch]

+++ Engine/Source/Editor/Sequencer/Private/SequencerCommonHelpers.cpp
@@ -55,7 +55,7 @@ void SequencerHelpers::GetDescendantNodes(TSharedRef<FSequencerDisplayNode> Disp
 
 int32 SequencerHelpers::TimeToFrame(float Time, float FrameRate)
 {
-	int32 Frame = Time * FrameRate;
+	float Frame = Time * FrameRate;
 	return FMath::RoundToInt(Frame);
 }
